+++ src/wine-host/editor.h
@@ -36,6 +36,17 @@
 #include "utils.h"
 
 /**
+ * The maximum number of Win32 messages to handle per message loop. This is
+ * needed because otherwise some plugins can run into an infinite loop. I've
+ * observed this with:
+ *
+ * - Waves plugins
+ * - Melda plugins when having multiple editor windows open within a single
+ *   plugin group
+ */
+constexpr int max_win32_messages = 20;
+
+/**
  * Used to store the maximum width and height of a screen.
  */
 struct Size {